requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pytest==7.4.0
pytest-cov==4.1.0